Code cleanup + multiversioning Notice: I didnt test forge version cuz dont have forge 1.19 installed
