+++ .architectury-transformer/debug.log
@@ -1,0 +1,1 @@
+[Architectury Transformer DEBUG] Closed File Systems for D:\mod\AllayInABottle\common\build\libs\Allay-In-A-Bottle-1.0.0.jar
