+++ .gitignore
@@ -31,3 +31,5 @@ bin/
 # fabric
 
 run/
+
+.architectury-transformer
