+++ common/src/main/resources/allaybottle-common.mixins.json
