+++ common/src/main/resources/allaybottle.accesswidener
@@ -1,0 +1,1 @@
+accessWidener   v2  named
