+++ common/src/main/resources/allaybottle.common.json
@@ -1,0 +1,3 @@
+{
+  "accessWidener": "allaybottle.accesswidener"
+}
