+++ fabric/src/main/kotlin/me/piguy/allaybottle/AllayBottleFabric.kt
@@ -1,0 +1,9 @@
+package me.piguy.allaybottle
+
+import net.fabricmc.api.ModInitializer
+
+class AllayBottleFabric : ModInitializer {
+    override fun onInitialize() {
+        AllayBottle.onInitialize()
+    }
+}
