+++ fabric/src/main/resources/assets/allaybottle/icon.png
