+++ fabric/src/main/resources/assets/allaybottle/lang/en_us.json
