+++ fabric/src/main/resources/assets/allaybottle/models/item/allay_bottle.json
