+++ fabric/src/main/resources/assets/allaybottle/textures/item/allay_bottle.png
