+++ forge/src/main/kotlin/me/piguy/allaybottle/AllayBottleForge.kt
@@ -1,0 +1,10 @@
+package me.piguy.allaybottle
+
+import net.minecraftforge.fml.common.Mod
+
+@Mod(AllayBottle.MOD_ID)
+class AllayBottleForge {
+    init {
+        AllayBottle.onInitialize()
+    }
+}
