+++ forge/src/main/resources/pack.mcmeta
@@ -1,0 +1,6 @@
+{
+  "pack": {
+    "description": "Allay In A Bottle",
+    "pack_format": 9
+  }
+}
