+++ gradle.properties
@@ -1,5 +1,5 @@
 kotlin.code.style = official
-org.gradle.jvmargs = -Xmx1G
+org.gradle.jvmargs = -Xmx3G
 org.gradle.warning.mode = all
 # Check these on https://fabricmc.net/develop/
 minecraftVersion = 1.19.2
@@ -8,10 +8,7 @@ loaderVersion = 0.14.9
 # Fabric API
 fabricVersion = 0.60.0+1.19.2
 loomVersion = 0.12-SNAPSHOT
-# Mod Properties
-modVersion = 1.0.0
-mavenGroup = me.piguy
-archivesBaseName = allay-in-a-bottle
 # Kotlin
-systemProp.kotlinVersion = 1.7.10
 fabricKotlinVersion = 1.8.3+kotlin.1.7.10
+# Forge
+forge_version=1.19.2-43.0.8
