--- src/main/kotlin/me/piguy/allaybottle/AllayBottleClient.kt
@@ -1,11 +1,0 @@
-package me.piguy.allaybottle
-
-import net.fabricmc.api.ClientModInitializer
-
-
-@Suppress("UNUSED")
-object AllayBottleClient: ClientModInitializer {
-  override fun onInitializeClient() {
-
-  }
-}
